+++ src/Config/YamlFileLoader.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace hanneskod\readmetester\Config;
+
+use Symfony\Component\Yaml\Yaml;
+
+final class YamlFileLoader
+{
+    private string $fileName;
+
+    public function __construct(string $fileName)
+    {
+        $this->fileName = $fileName;
+    }
+
+    public function loadYamlFile(ConfigManager $manager): void
+    {
+        if (is_readable($this->fileName)) {
+            // TODO dispatch log event that file is loaded
+            $manager->loadRepository(
+                new ArrayRepository(Yaml::parseFile($this->fileName))
+            );
+        }
+    }
+}