+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project now adheres to [Semantic Versioning](http://semver.org/).
 - Added the `--ignore` command line option
 - Added the `--stop-on-failure` command line option
 - Using php8 attributes instead of annotations
+- Added support for `readme-tester.yaml` config file
 
 ## [1.0.0] - 2020-07-06